SuperAgent-Backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-binance
web3
eth-account
